refactor: refactor spi class loader
