+++ sdk-component/src/main/resources/capa-component-configuration.properties
@@ -1,0 +1,2 @@
+# configuration component
+group.rxcloud.capa.component.configstore.CapaConfigStore=your-spi-class-path
