+++ sdk-component/src/main/resources/capa-component-pubsub.properties
@@ -1,0 +1,2 @@
+# pubsub component
+group.rxcloud.capa.component.pubsub.CapaPubSub=your-spi-class-path
