+++ sdk-component/src/main/resources/capa-component-rpc.properties
@@ -1,0 +1,2 @@
+# rpc component
+group.rxcloud.capa.component.http.CapaHttp=your-spi-class-path
