+++ sdk-spi-demo/src/main/java/group/rxcloud/capa/spi/demo/http/DemoCapaHttp.java
@@ -18,8 +18,8 @@ package group.rxcloud.capa.spi.demo.http;
 
 import group.rxcloud.capa.component.http.HttpResponse;
 import group.rxcloud.capa.infrastructure.serializer.CapaObjectSerializer;
-import group.rxcloud.capa.spi.config.RpcServiceOptions;
-import group.rxcloud.capa.spi.demo.config.DemoRpcServiceOptions;
+import group.rxcloud.capa.spi.http.config.RpcServiceOptions;
+import group.rxcloud.capa.spi.demo.http.config.DemoRpcServiceOptions;
 import group.rxcloud.capa.spi.http.CapaSerializeHttpSpi;
 import group.rxcloud.cloudruntimes.utils.TypeRef;
 import okhttp3.OkHttpClient;
