+++ sdk-spi-demo/src/main/resources/capa-component-configuration.properties
@@ -1,1 +1,2 @@
+# configuration component
 group.rxcloud.capa.component.configstore.CapaConfigStore=group.rxcloud.capa.spi.demo.configstore.DemoCapaConfigStore
