+++ sdk-spi-demo/src/main/resources/capa-component-pubsub.properties
@@ -1,1 +1,2 @@
+# pubsub component
 group.rxcloud.capa.component.pubsub.CapaPubSub=group.rxcloud.capa.spi.demo.pubsub.DemoCapaPubSub
