+++ sdk-spi-demo/src/main/resources/capa-component-rpc.properties
@@ -1,2 +1,2 @@
+# rpc component
 group.rxcloud.capa.component.http.CapaHttp=group.rxcloud.capa.spi.demo.http.DemoCapaHttp
-group.rxcloud.capa.spi.config.CapaSpiOptionsLoader=group.rxcloud.capa.spi.demo.config.DemoSpiOptionsLoader
