+++ sdk-spi-demo/src/main/resources/capa-infrastructure-cloud.properties
@@ -1,0 +1,2 @@
+# {@code CapaEnvironment} spi class load
+group.rxcloud.capa.infrastructure.CapaEnvironment=group.rxcloud.capa.spi.demo.DemoCapaEnvironment
