+++ sdk-spi/src/main/java/group/rxcloud/capa/spi/http/CapaHttpSpi.java
@@ -19,9 +19,9 @@ package group.rxcloud.capa.spi.http;
 import group.rxcloud.capa.component.http.CapaHttp;
 import group.rxcloud.capa.component.http.HttpResponse;
 import group.rxcloud.capa.infrastructure.serializer.CapaObjectSerializer;
-import group.rxcloud.capa.spi.config.CapaSpiOptionsLoader;
-import group.rxcloud.capa.spi.config.CapaSpiProperties;
-import group.rxcloud.capa.spi.config.RpcServiceOptions;
+import group.rxcloud.capa.spi.http.config.CapaSpiOptionsLoader;
+import group.rxcloud.capa.spi.http.config.CapaSpiProperties;
+import group.rxcloud.capa.spi.http.config.RpcServiceOptions;
 import group.rxcloud.cloudruntimes.domain.core.invocation.HttpExtension;
 import group.rxcloud.cloudruntimes.utils.TypeRef;
 import okhttp3.OkHttpClient;
