+++ sdk-spi/src/main/java/group/rxcloud/capa/spi/http/config/RpcServiceOptions.java
@@ -14,26 +14,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package group.rxcloud.capa.spi.demo.config;
 
-import group.rxcloud.capa.spi.config.RpcServiceOptions;
+package group.rxcloud.capa.spi.http.config;
+
 
 /**
  * RPC service options. Define for AppId.
  */
-public class DemoRpcServiceOptions implements RpcServiceOptions {
-
-    /**
-     * Unique rpc service ID
-     */
-    private final String appId;
-
-    /**
-     * Instantiates a new Capa rpc service options.
-     *
-     * @param appId the app id
-     */
-    public DemoRpcServiceOptions(String appId) {
-        this.appId = appId;
-    }
+public interface RpcServiceOptions {
 }
