+++ sdk-spi/src/main/resources/capa-component-rpc-common.properties
@@ -1,0 +1,2 @@
+# rpc spi options class loader
+group.rxcloud.capa.spi.http.config.CapaSpiOptionsLoader=your-spi-class-path
