+++ sdk-spi/src/test/resources/capa-component-rpc.properties
@@ -1,1 +1,1 @@
-group.rxcloud.capa.spi.config.CapaSpiOptionsLoader=group.rxcloud.capa.spi.config.TestSpiOptionsLoader
+group.rxcloud.capa.spi.http.config.CapaSpiOptionsLoader=group.rxcloud.capa.spi.config.TestSpiOptionsLoader
